CSI - Surfaces klog flags; common tests aggregation
