+++ pkg/volume/csi/csi_attacher_test.go
@@ -334,9 +334,9 @@ func TestAttacherWithCSIDriver(t *testing.T) {
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			fakeClient := fakeclient.NewSimpleClientset(
-				getCSIDriver("not-attachable", nil, &bFalse),
-				getCSIDriver("attachable", nil, &bTrue),
-				getCSIDriver("nil", nil, nil),
+				getTestCSIDriver("not-attachable", nil, &bFalse),
+				getTestCSIDriver("attachable", nil, &bTrue),
+				getTestCSIDriver("nil", nil, nil),
 			)
 			plug, fakeWatcher, tmpDir, _ := newTestWatchPlugin(t, fakeClient)
 			defer os.RemoveAll(tmpDir)
@@ -418,9 +418,9 @@ func TestAttacherWaitForVolumeAttachmentWithCSIDriver(t *testing.T) {
 	for _, test := range tests {
 		t.Run(test.name, func(t *testing.T) {
 			fakeClient := fakeclient.NewSimpleClientset(
-				getCSIDriver("not-attachable", nil, &bFalse),
-				getCSIDriver("attachable", nil, &bTrue),
-				getCSIDriver("nil", nil, nil),
+				getTestCSIDriver("not-attachable", nil, &bFalse),
+				getTestCSIDriver("attachable", nil, &bTrue),
+				getTestCSIDriver("nil", nil, nil),
 			)
 			plug, tmpDir := newTestPlugin(t, fakeClient)
 			defer os.RemoveAll(tmpDir)
