+++ pkg/volume/csi/csi_mounter_test.go
@@ -17,10 +17,7 @@ limitations under the License.
 package csi
 
 import (
-	"bytes"
-	"encoding/json"
 	"fmt"
-	"io/ioutil"
 	"math/rand"
 	"os"
 	"path"
@@ -30,7 +27,6 @@ import (
 
 	api "k8s.io/api/core/v1"
 	storage "k8s.io/api/storage/v1"
-	storagev1beta1 "k8s.io/api/storage/v1beta1"
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	utilfeature "k8s.io/apiserver/pkg/util/feature"
@@ -147,9 +143,9 @@ func MounterSetUpTests(t *testing.T, podInfoEnabled bool) {
 		t.Run(test.name, func(t *testing.T) {
 			klog.Infof("Starting test %s", test.name)
 			fakeClient := fakeclient.NewSimpleClientset(
-				getCSIDriver("no-info", &noPodMountInfo, nil),
-				getCSIDriver("info", &currentPodInfoMount, nil),
-				getCSIDriver("nil", nil, nil),
+				getTestCSIDriver("no-info", &noPodMountInfo, nil),
+				getTestCSIDriver("info", &currentPodInfoMount, nil),
+				getTestCSIDriver("nil", nil, nil),
 			)
 			plug, tmpDir := newTestPlugin(t, fakeClient)
 			defer os.RemoveAll(tmpDir)
@@ -694,64 +690,4 @@ func TestUnmounterTeardown(t *testing.T) {
 		t.Error("csi server may not have received NodeUnpublishVolume call")
 	}
 
-}
-
-func TestSaveVolumeData(t *testing.T) {
-	plug, tmpDir := newTestPlugin(t, nil)
-	defer os.RemoveAll(tmpDir)
-	testCases := []struct {
-		name       string
-		data       map[string]string
-		shouldFail bool
-	}{
-		{name: "test with data ok", data: map[string]string{"key0": "val0", "_key1": "val1", "key2": "val2"}},
-		{name: "test with data ok 2 ", data: map[string]string{"_key0_": "val0", "&key1": "val1", "key2": "val2"}},
-	}
-
-	for i, tc := range testCases {
-		t.Logf("test case: %s", tc.name)
-		specVolID := fmt.Sprintf("spec-volid-%d", i)
-		mountDir := path.Join(getTargetPath(testPodUID, specVolID, plug.host), "/mount")
-		if err := os.MkdirAll(mountDir, 0755); err != nil && !os.IsNotExist(err) {
-			t.Errorf("failed to create dir [%s]: %v", mountDir, err)
-		}
-
-		err := saveVolumeData(path.Dir(mountDir), volDataFileName, tc.data)
-
-		if !tc.shouldFail && err != nil {
-			t.Errorf("unexpected failure: %v", err)
-		}
-		// did file get created
-		dataDir := getTargetPath(testPodUID, specVolID, plug.host)
-		file := path.Join(dataDir, volDataFileName)
-		if _, err := os.Stat(file); err != nil {
-			t.Errorf("failed to create data dir: %v", err)
-		}
-
-		// validate content
-		data, err := ioutil.ReadFile(file)
-		if !tc.shouldFail && err != nil {
-			t.Errorf("failed to read data file: %v", err)
-		}
-
-		jsonData := new(bytes.Buffer)
-		if err := json.NewEncoder(jsonData).Encode(tc.data); err != nil {
-			t.Errorf("failed to encode json: %v", err)
-		}
-		if string(data) != jsonData.String() {
-			t.Errorf("expecting encoded data %v, got %v", string(data), jsonData)
-		}
-	}
-}
-
-func getCSIDriver(name string, podInfoMount *bool, attachable *bool) *storagev1beta1.CSIDriver {
-	return &storagev1beta1.CSIDriver{
-		ObjectMeta: meta.ObjectMeta{
-			Name: name,
-		},
-		Spec: storagev1beta1.CSIDriverSpec{
-			PodInfoOnMount: podInfoMount,
-			AttachRequired: attachable,
-		},
-	}
 }
